long-term-memory/requirements.txt
langgraph
langchain
langchain-openai
langchain-community
psycopg2-binary
python-dotenv
sqlalchemy
sentence-transformers
pgvector
numpy
